requirements.txt
asgiref==3.11.0
contourpy==1.3.2
cycler==0.12.1
Django==5.2.8
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
fonttools==4.60.1
joblib==1.5.2
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.2.6
packaging==25.0
pillow==12.0.0
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
typing_extensions==4.15.0
